requirements.txt
alabaster==0.7.16
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.10
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
Pygments==2.18.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
